leetcode/0494.Target-Sum/494.TargetSum_test.go
package leetcode

import "testing"

func Test_findTargetSumWays(t *testing.T) {
	type args struct {
		nums   []int
		target int
	}
	tests := []struct {
		name string
		args args
		want int
	}{
		{
			name: "nums = [1,1,1,1,1], target = 3",
			args: args{[]int{1, 1, 1, 1, 1}, 3},
			want: 5,
		},
		{
			name: "nums = [1], target = 1",
			args: args{[]int{1}, 1},
			want: 1,
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := findTargetSumWays(tt.args.nums, tt.args.target); got != tt.want {
				t.Errorf("findTargetSumWays() = %v, want %v", got, tt.want)
			}
			if got := findTargetSumWays1(tt.args.nums, tt.args.target); got != tt.want {
				t.Errorf("findTargetSumWays() = %v, want %v", got, tt.want)
			}
		})
	}
}
